Fixes #4, removed unneeded comments, re factored letter calculation to its own method
Test data was not created, but I am reasonably sure that I have fixed the problem. If not, I will reopen the issue and do this the proper way. The issue was if the weighted grade was, as an example, 89.5, it is technically larger than 89, so it would fall through all the blocks and be counted as an F. The fix is to truncate the weighted grade by typecasting it to an int.